+++ lib/models/section_model.dart
@@ -107,7 +107,7 @@ class SectionModel extends ChangeNotifier {
     }
 
     for(final original in originalItems){
-      if(!items.contains(original)){
+      if(!items.contains(original) && (original.image as String).contains('firebase')){
         try {
           final ref = storage.refFromURL(original.image);
           await ref.delete();
@@ -128,11 +128,13 @@ class SectionModel extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for(final item in items){
-      try {
-        final ref = storage.refFromURL(item.image as String);
-        await ref.delete();
-      } catch (e){
-        logger.e('Error ao remover a imagem do firestorm');
+      if((item.image as String).contains('firebase')){
+        try {
+          final ref = storage.refFromURL(item.image as String);
+          await ref.delete();
+        } catch (e){
+          logger.e('Error ao remover a imagem do firestorm');
+        }
       }
     }
   }
